+++ ios/Classes/src/features/render/RenderVideo.swift
@@ -74,7 +74,8 @@ class RenderVideo {
                                 inputPath: newPath,
                                 startUs: clip.startUs,
                                 endUs: clip.endUs,
-                                volume: clip.volume
+                                volume: clip.volume,
+                                playbackSpeed: clip.playbackSpeed
                             )
                         }
                         return clip